refactor: move remaining event listeners to separate components
