+++ packages/api-client/src/components/ImportCollection/DropEventListener.vue
@@ -1,0 +1,116 @@
+<script lang="ts" setup>
+import { ScalarIcon } from '@scalar/components'
+import { onBeforeUnmount, onMounted, ref } from 'vue'
+
+const emit = defineEmits<{
+  (e: 'input', value: string): void
+}>()
+
+const isDragging = ref<boolean>(false)
+
+let dragCounter = 0
+
+// Register event listeners
+onMounted(() => {
+  document.addEventListener('dragenter', handleDragEnter)
+  document.addEventListener('dragleave', handleDragLeave)
+  document.addEventListener('dragover', handleDragOver)
+  document.addEventListener('drop', handleDrop)
+})
+
+// Unregister event listeners
+onBeforeUnmount(() => {
+  document.removeEventListener('dragenter', handleDragEnter)
+  document.removeEventListener('dragover', handleDragOver)
+  document.removeEventListener('dragleave', handleDragLeave)
+  document.removeEventListener('drop', handleDrop)
+})
+
+// Drop
+async function handleDrop(event: DragEvent) {
+  event.preventDefault()
+  isDragging.value = false
+  dragCounter = 0
+
+  if (event.dataTransfer) {
+    // Text
+    const droppedText = event.dataTransfer.getData('text').replace(/^blob:/, '')
+
+    if (droppedText) {
+      emit('input', droppedText)
+    }
+    // Files
+    else if (event.dataTransfer.files.length > 0) {
+      const file = event.dataTransfer.files[0]
+      const reader = new FileReader()
+
+      reader.onload = async (e) => {
+        if (e.target && typeof e.target.result === 'string') {
+          emit('input', e.target.result)
+        }
+      }
+      reader.readAsText(file)
+    }
+  }
+}
+
+function handleDragOver(event: DragEvent) {
+  event.preventDefault()
+}
+
+function handleDragLeave(event: DragEvent) {
+  event.preventDefault()
+  dragCounter--
+
+  if (dragCounter === 0) {
+    isDragging.value = false
+  }
+}
+
+function handleDragEnter(event: DragEvent) {
+  event.preventDefault()
+  dragCounter++
+
+  if (event.dataTransfer) {
+    const items = event.dataTransfer.items
+    for (let i = 0; i < items.length; i++) {
+      const item = items[i]
+
+      if (
+        item.kind === 'string' ||
+        item.type.includes('json') ||
+        item.type.includes('yml') ||
+        item.type.includes('yaml')
+      ) {
+        isDragging.value = true
+        return
+      }
+    }
+  }
+  isDragging.value = false
+}
+</script>
+
+<template>
+  <transition
+    enterActiveClass="transition-opacity duration-200"
+    enterFromClass="opacity-0"
+    leaveActiveClass="transition-opacity duration-200"
+    leaveToClass="opacity-0">
+    <div
+      v-if="isDragging"
+      class="fixed bottom-10 right-10 w-64 h-64 bg-b-2 z-50 rounded border transition-opacity duration-200">
+      <div class="flex flex-col items-center justify-center h-full">
+        <div>
+          <ScalarIcon
+            icon="Download"
+            size="md"
+            thickness="1.75" />
+        </div>
+        <div class="text-center m-4 text-c-1">
+          Drop your OpenAPI document here
+        </div>
+      </div>
+    </div>
+  </transition>
+</template>
